+++ app/users/presentation/__init__.py
@@ -1,5 +1,0 @@
-from app.users.presentation.routes import router as user_router
-
-__all__ = [
-    "user_router",
-]